+++ CopterMotor.cpp
@@ -14,7 +14,7 @@ void CopterMotor::invoke(int _power)
 {
 	QString s;
 	m_ctrlFile.open(QIODevice::WriteOnly|QIODevice::Truncate|QIODevice::Unbuffered|QIODevice::Text);
-	double powerFactor = (double)(m_powerMax - m_powerMin) / 100;
+	float powerFactor = (float)(m_powerMax - m_powerMin) / 100;
 	s.sprintf("%d\n", static_cast<int>(_power * powerFactor + m_powerMin));
 	m_ctrlFile.write(s.toLatin1());
 	m_ctrlFile.close();
@@ -25,8 +25,8 @@ CopterMotor::CopterMotor(QSettings* settings, const QString& _ctrlPath) :
 	m_ctrlFile(_ctrlPath),
 	m_delta(1.0)
 {
-	m_powerMin = m_settings->value("MotorMin").toDouble();
-	m_powerMax = m_settings->value("MotorMax").toDouble();
+	m_powerMin = m_settings->value("MotorMin").toFloat();
+	m_powerMax = m_settings->value("MotorMax").toFloat();
 
 	invoke_open();
 }
@@ -36,18 +36,18 @@ CopterMotor::~CopterMotor()
 	invoke_close();
 }
 
-void CopterMotor::delta(double _delta)
+void CopterMotor::delta(float _delta)
 {
-	m_delta = qMax(qMin(_delta, 100.0), -100.0);
+	m_delta = qMax(qMin(_delta, 100.0f), -100.0f);
 }
 
 void CopterMotor::setPower(unsigned _power)
 {
-	m_delta = qMax(m_delta, - static_cast<double>(_power));
+	m_delta = qMax(m_delta, - static_cast<float>(_power));
 	int pwr =  floor(_power + sqrt(m_delta + 0.5));
 	pwr = qMin(pwr, static_cast<int>(_power) * 2);
 	invoke(pwr);
-	emit powerChanged(static_cast<double>(pwr));
+	emit powerChanged(static_cast<float>(pwr));
 }
