+++ CopterMotor.hpp
@@ -11,13 +11,13 @@ public:
 	CopterMotor(QSettings* settings, const QString& _ctrlPath);
 	~CopterMotor();
 
-	double delta() const { return m_delta; }
-	void delta(double _factor);
+	float delta() const { return m_delta; }
+	void delta(float _factor);
 
 	void setPower(unsigned _power);
 
 signals:
-	void powerChanged(double power);
+	void powerChanged(float power);
 
 public slots:
 	void emergencyStop();
@@ -25,8 +25,8 @@ public slots:
 protected:
 	QSettings* m_settings;
 	QFile       m_ctrlFile;
-	double      m_delta;
-	double m_powerMax, m_powerMin; // real power, to write to ctrlFile
+	float      m_delta;
+	float m_powerMax, m_powerMin; // real power, to write to ctrlFile
 	void invoke_open();
 	void invoke_close();
 	void invoke(int _power);
